update: refactor table panel
