+++ Library_System/.metadata/.gitignore
@@ -1,0 +1,2 @@
+/.lock
+/version.ini
