+++ Library_System/.metadata/.plugins/org.eclipse.core.resources/.root/.gitignore
@@ -1,0 +1,1 @@
+/1.tree
