+++ Library_System/.metadata/.plugins/org.eclipse.core.resources/.root/.indexes/.gitignore
@@ -1,0 +1,3 @@
+/history.version
+/properties.index
+/properties.version
