+++ Library_System/.metadata/.plugins/org.eclipse.core.resources/.safetable/.gitignore
@@ -1,0 +1,1 @@
+/org.eclipse.core.resources
