+++ Library_System/.metadata/.plugins/org.eclipse.e4.workbench/.gitignore
@@ -1,0 +1,1 @@
+/workbench.xmi
