+++ Library_System/.metadata/.plugins/org.eclipse.jdt.core/.gitignore
@@ -1,0 +1,4 @@
+/assumedExternalFilesCache
+/externalFilesCache
+/nonChainingJarsCache
+/variablesAndContainers.dat
