+++ Library_System/.metadata/.plugins/org.eclipse.jdt.ui/.gitignore
@@ -1,0 +1,2 @@
+/OpenTypeHistory.xml
+/QualifiedTypeNameHistory.xml
