+++ Library_System/.metadata/.plugins/org.eclipse.m2e.logback.configuration/.gitignore
@@ -1,0 +1,1 @@
+/logback.1.16.0.20200318-1040.xml
