+++ Library_System/.metadata/.plugins/org.eclipse.oomph.setup/.gitignore
@@ -1,0 +1,1 @@
+/workspace.setup
