+++ Library_System/.metadata/.plugins/org.eclipse.rse.core/initializerMarks/.gitignore
@@ -1,0 +1,1 @@
+/org.eclipse.rse.internal.core.RSELocalConnectionInitializer.mark
