+++ Library_System/.metadata/.plugins/org.eclipse.tips.ide/.gitignore
@@ -1,0 +1,1 @@
+/dialog_settings.xml
