+++ Library_System/.metadata/.plugins/org.eclipse.tm.terminal.view.ui/.executables/.gitignore
@@ -1,0 +1,1 @@
+/data.properties
