+++ Library_System/.metadata/.plugins/org.eclipse.ui.intro/.gitignore
@@ -1,0 +1,1 @@
+/introstate
