+++ Library_System/.metadata/.plugins/org.eclipse.ui.workbench/.gitignore
@@ -1,0 +1,1 @@
+/workingsets.xml
