+++ Library_System/src/view/manager/ManagerPanel.java
@@ -3,6 +3,7 @@ package view.manager;
 import javax.swing.JPanel;
 
 import view.component.DefaultPanel;
+import view.component.DefaultSearchedTablePanel;
 
 public class ManagerPanel extends DefaultPanel{
 	
@@ -10,7 +11,7 @@ public class ManagerPanel extends DefaultPanel{
 		super();
 		
 		JPanel title = new TitlePanel();
-		JPanel table = new TablePanel();
+		JPanel table = new DefaultSearchedTablePanel();
 		
 		add(title);
 	  	add(table);
